moved pdf builder outside of plot directory
